+++ packages/shared/src/index.ts
@@ -9,3 +9,4 @@ export * from "./api_types";
 export * from "./grid";
 export * from "./coordinate";
 export * from "./time_control";
+export * from "./utils";
