+++ packages/shared/src/utils.ts
@@ -1,0 +1,16 @@
+import { MovesType } from "./abstract_game";
+
+export function getOnlyMove(moves: MovesType): {
+  player: number;
+  move: string;
+} {
+  const players = Object.keys(moves);
+  if (players.length > 1) {
+    throw new Error(`More than one player: ${players}`);
+  }
+  if (players.length === 0) {
+    throw new Error("No players specified!");
+  }
+  const player = Number(players[0]);
+  return { player, move: moves[player] };
+}
